server/requirements.txt
Flask
flask-cors
torch
Pillow
